fix: Wanna know? Nevermind
Removed the subshell that created an array with all the variables in the script. Why? Because it was creating an array with all variables defined in the enviroment, not only from the script.

The error handling functions will now use an empty array, which doesn't in fact affects its job of error handling.